[feat] Separate the kernels from the main estimator.
