+++ emir/estimators/kernels/gaussian_kernel.py
@@ -1,0 +1,119 @@
+"""
+Courtesy of https://github.com/g-pichler/knife/blob/master/da_experiments/estimators/knife.py
+From https://arxiv.org/abs/2202.06618
+"""
+
+import numpy as np
+import torch
+import torch.nn as nn
+
+from .kernels import BaseMargKernel, BaseCondKernel
+from .feed_forward import FF
+
+
+class GaussianMargKernel(BaseMargKernel):
+    """
+    Used to compute p(z_d)
+    """
+
+    def __init__(self, args, zc_dim, zd_dim, init_samples=None, **kwargs):
+        super().__init__(args, zc_dim, zd_dim)
+        self.K = args.marg_modes if self.optimize_mu else args.batch_size
+        self.init_std = args.init_std
+
+        self.logC = torch.tensor([-self.d / 2 * np.log(2 * np.pi)])
+
+        if init_samples is None:
+            init_samples = self.init_std * torch.randn((self.K, self.d))
+        # self.means = nn.Parameter(torch.rand(self.K, self.d), requires_grad=True)
+        if self.optimize_mu:
+            self.means = nn.Parameter(init_samples, requires_grad=True)  # [K, db]
+        else:
+            self.means = nn.Parameter(init_samples, requires_grad=False)
+
+        if args.cov_diagonal == "var":
+            diag = self.init_std * torch.randn((1, self.K, self.d))
+        else:
+            diag = self.init_std * torch.randn((1, 1, self.d))
+        self.logvar = nn.Parameter(diag, requires_grad=True)
+
+        if args.cov_off_diagonal == "var":
+            tri = self.init_std * torch.randn((1, self.K, self.d, self.d))
+            tri = tri.to(init_samples.dtype)
+            self.tri = nn.Parameter(tri, requires_grad=True)
+        else:
+            self.tri = None
+
+        weigh = torch.ones((1, self.K))
+        if args.average == "var":
+            self.weigh = nn.Parameter(weigh, requires_grad=True)
+        else:
+            self.weigh = nn.Parameter(weigh, requires_grad=False)
+
+    def logpdf(self, x):
+        assert len(x.shape) == 2 and x.shape[1] == self.d, "x has to have shape [N, d]"
+        x = x[:, None, :]
+        w = torch.log_softmax(self.weigh, dim=1)
+        y = x - self.means
+        logvar = self.logvar
+        if self.use_tanh:
+            logvar = logvar.tanh()
+        var = logvar.exp()
+        y = y * var
+        # print(f"Marg : {var.min()} | {var.max()} | {var.mean()}")
+        if self.tri is not None:
+            y = y + torch.squeeze(
+                torch.matmul(torch.tril(self.tri, diagonal=-1), y[:, :, :, None]), 3
+            )
+        y = torch.sum(y**2, dim=2)
+
+        y = -y / 2 + torch.sum(torch.log(torch.abs(var) + 1e-8), dim=-1) + w
+        y = torch.logsumexp(y, dim=-1)
+        return self.logC.to(y.device) + y
+
+    def update_parameters(self, z):
+        self.means = z
+
+
+class GaussianCondKernel(BaseCondKernel):
+    """
+    Used to compute p(z_d | z_c)
+    """
+
+    def __init__(self, args, zc_dim, zd_dim, **kwargs):
+        super().__init__(args, zc_dim, zd_dim)
+        self.K = args.cond_modes
+        self.logC = torch.tensor([-self.d / 2 * np.log(2 * np.pi)])
+
+        self.mu = FF(args, zc_dim, self.d, self.K * zd_dim)
+        self.logvar = FF(args, zc_dim, self.d, self.K * zd_dim)
+
+        self.weight = FF(args, zc_dim, self.d, self.K)
+        self.tri = None
+        if args.cov_off_diagonal == "var":
+            self.tri = FF(args, zc_dim, self.d, self.K * zd_dim**2)
+
+    def logpdf(self, z_c, z_d):  # H(z_d|z_c)
+        z_d = z_d[:, None, :]  # [N, 1, d]
+
+        w = torch.log_softmax(self.weight(z_c), dim=-1)  # [N, K]
+        mu = self.mu(z_c)
+        logvar = self.logvar(z_c)
+        if self.use_tanh:
+            logvar = logvar.tanh()
+        var = logvar.exp().reshape(-1, self.K, self.d)
+        mu = mu.reshape(-1, self.K, self.d)
+        # print(f"Cond : {var.min()} | {var.max()} | {var.mean()}")
+
+        z = z_d - mu  # [N, K, d]
+        z = var * z
+        if self.tri is not None:
+            tri = self.tri(z_c).reshape(-1, self.K, self.d, self.d)
+            z = z + torch.squeeze(
+                torch.matmul(torch.tril(tri, diagonal=-1), z[:, :, :, None]), 3
+            )
+        z = torch.sum(z**2, dim=-1)  # [N, K]
+
+        z = -z / 2 + torch.log(torch.abs(var) + 1e-8).sum(-1) + w
+        z = torch.logsumexp(z, dim=-1)
+        return self.logC.to(z.device) + z
